style split data_check to two modules
