+++ 3750/2/commands/checkin.rb
@@ -1,20 +1,10 @@
-require_relative '../data_check'
-
 module Commands
   module Checkin
-    include DataCheck
-
     def checkin!(*)
       notify(:not_registered) && return unless registered?
-      notify(:not_checkout) && return unless checkout?
+      notify(:not_checked_out) && return unless checked_out?
 
       process_checkin
-      notify(:success_check_start)
-    end
-
-    def checkin_ending
-      notify(:success_checkin_end)
-      set_checkin_flags
     end
 
     private
@@ -22,12 +12,8 @@ module Commands
     def process_checkin
       session[:command] = 'checkin'
       session[:timestamp] = Time.now.getutc.to_i
-      save_context :photo_check
-    end
-
-    def set_checkin_flags
-      session[:checkin?] = true
-      session[:checkout?] = false
+      notify(:success_check_start)
+      save_context(:photo_check)
     end
   end
 end
