+++ 3750/2/commands/checkout.rb
@@ -2,16 +2,9 @@ module Commands
   module Checkout
     def checkout!(*)
       notify(:not_registered) && return unless registered?
-      notify(:not_checkin) && return unless checkin?
+      notify(:not_checked_in) && return unless checked_in?
 
       process_checkout
-      notify(:success_check_start)
-    end
-
-    def checkout_ending
-      notify(:success_checkout_end)
-      send_sticker(:farewell_sticker)
-      set_checkout_flags
     end
 
     private
@@ -19,12 +12,8 @@ module Commands
     def process_checkout
       session[:command] = 'checkout'
       session[:timestamp] = Time.now.getutc.to_i
-      save_context :photo_check
-    end
-
-    def set_checkout_flags
-      session[:checkin?] = false
-      session[:checkout?] = true
+      notify(:success_check_start)
+      save_context(:photo_check)
     end
   end
 end
